+++ src/index.css
@@ -1514,40 +1514,157 @@ button:focus-visible {
   color: black;
 }
 
-/* Estilos CSS personalizados */
-
-
-.registroform {
-  margin-bottom: 10px;
-}
-
-.registroform label {
-  display: block;
-  margin-bottom: 5px;
-  color: white; /* Texto negro */
-}
-
-.registroform input {
-  width: calc(100% - 80px); /* Ancho reducido */
-  padding: 8px;
-  border: 1px solid #ccc; /* Borde gris */
-  border-radius: 3px;
-}
-
-.button-registro {
-  background-color: #333; /* Fondo negro */
-  color: #fff; /* Texto blanco */
-  border: none;
-  padding: 10px 20px;
-  border-radius: 3px;
-  cursor: pointer;
-}
-
-.button-registro:hover {
-  background-color: #555; /* Fondo gris oscuro al pasar el mouse */
-}
-
 .required {
   color: white;
   margin-left: 5px;
 }
+
+/* ESTILOS PARA EL FORMULARIO DE LOGUEO DE USUARIO */
+
+.login-container {
+  background-color: #fff;
+  padding: 30px; 
+  padding-top: 5px;
+  border-radius: 16px; 
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.2); 
+  border: 2px solid #ccc; 
+  max-width: 1200px;
+  margin: auto;
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  z-index: 2000;
+ 
+}
+
+.button-close-login {
+  background: none;
+  border: none;
+  color: #000;
+  font-size: 16px;
+  cursor: pointer;
+  display: flex;
+  margin-bottom: 20px;
+  justify-content: flex-end;
+}
+
+
+
+.login-form {
+  display: flex;
+  flex-direction: column;
+}
+
+.login-form .form-group {
+  margin-bottom: 16px; 
+}
+
+.login-form label {
+  color: black;
+  margin-bottom: 8px; 
+  font-weight: 600;
+}
+
+.login-form input[type="email"],
+.login-form input[type="text"] {
+  width: calc(100% - 16px); 
+  padding: 8px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  margin-right: 8px; 
+}
+
+.login-form input[type="submit"] {
+  width: 100%;
+  padding: 12px; 
+  color: #ffffff;
+  text-shadow: 0 -1px 0 rgba(0, 0, 0, 0.25);
+  background-color: #5bb75b;
+  background-image: linear-gradient(to bottom, #62c462, #51a351);
+  background-repeat: repeat-x;
+  border: 1px solid #387038;
+  border-radius: 4px;
+  cursor: pointer;
+  margin-top: 20px; 
+}
+
+.login-form input[type="submit"]:hover {
+  background-color: #51a351;
+}
+
+
+/* ESTILOS PARA EL FORMULARIO DE REGISTRO DE USUARIO */
+
+.registro-container {
+  background-color: #fff;
+  padding: 30px;
+  padding-top: 10px;
+  border-radius: 16px;
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.2);
+  border: 2px solid #ccc;
+  max-width: 600px;
+  z-index: 2000;
+  margin: auto;
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  
+}
+
+.registro {
+  display: flex;
+  flex-wrap: wrap;
+}
+
+.columna {
+  width: 50%; 
+}
+
+.registroform {
+  width: 100%; 
+  margin-bottom: 16px; 
+}
+
+.registroform label {
+  color: black;
+  font-weight: 600;
+}
+
+.registroform input {
+  width: calc(90% - 16px); 
+  padding: 8px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+}
+
+.button-registro {
+  width: calc(100% - 16px); 
+  padding: 12px;
+  color: #ffffff;
+  text-shadow: 0 -1px 0 rgba(0, 0, 0, 0.25);
+  background-color: #5bb75b;
+  background-image: linear-gradient(to bottom, #62c462, #51a351);
+  background-repeat: repeat-x;
+  border: 1px solid #387038;
+  border-radius: 4px;
+  cursor: pointer;
+  margin-top: 20px;
+}
+
+.button-registro:hover {
+  background-color: #51a351;
+}
+
+.registro-container span {
+  color: blue; 
+ 
+}
+
+
+
+
+
+
+
